+++ packages/vital-templates/src/components/GenericTemplate/tokens/vitalGenericTemplate.ts
@@ -52,7 +52,7 @@ const WithNoBreadcrumbsOnHomePage = asGenericTemplateToken({
   },
 });
 
-const Base = asGenericTemplateToken({
+const Default = asGenericTemplateToken({
   Components: {
     PageWrapper: vitalLayout.Default,
     Breadcrumb: as(vitalBreadcrumbs.Default),
@@ -94,36 +94,40 @@ const Base = asGenericTemplateToken({
   Compose: {
     WithNoBreadcrumbsOnHomePage,
   },
+  Meta: {
+    title: 'Default',
+  },
 });
 
 const ContentListing = asGenericTemplateToken({
-  ...Base,
+  ...Default,
   Meta: {
     title: 'Content Listing',
   },
   Components: {
-    ...Base.Components,
+    ...Default.Components,
     Content: on(ContentListingClean)(vitalContentListing.Default),
   },
   Schema: {
-    ...Base.Schema,
+    ...Default.Schema,
     Content: as(
       withNodeKey({ nodeKey: 'content-listing', nodeCollection: 'site' }),
       withNode,
-      Base.Schema.Content,
+      Default.Schema.Content,
     ),
   }
 });
 
+// Holding here for backward compabitility document as deprecated.
 const Generic = asGenericTemplateToken({
-  ...Base,
+  ...Default,
   Meta: {
     title: 'Generic',
   },
 });
 
 export default {
-  Base,
+  Default,
   Generic,
   ContentListing,
   WithNoBreadcrumbsOnHomePage,