+++ packages/vital-test/src/shadow/@bodiless/vital-layout/Helmet.ts
@@ -12,13 +12,16 @@
  * limitations under the License.
  */
 
+// Note: when used with the __vital__ package on vital-demo, that package takes precedence
+// over this one.  This test override has no effect and just an example.
+
 import { vitalHelmetBase } from '@bodiless/vital-layout';
 import { asFluidToken } from '@bodiless/vital-elements';
 import { addProps } from '@bodiless/fclasses';
 
 const Default = asFluidToken(vitalHelmetBase.Default, {
   Behavior: {
-    BodyHelmet: addProps({ 'data-shadowed-by': '__vital__:Helmet' }),
+    BodyHelmet: addProps({ 'data-shadowed-by': '__vitaltest__:Helmet' }),
   },
 });